this is a practice using a integrated LSH lib
this algorithm use a array(bxr) of hyper planes to determine whether the point is labeled the plus or minus, and get the “probability” by apply and-or on the results. I used the 28x28 dimension handwriting picture pixels data for test. it seems with small matrix of hyperplanes(bxr), it cannot cluster the sparse vectors and if we raise the (bxr), it takes a long time to deal with the data.
the zip file includes the data in csv files and LSH_test.py is the main function. the LSH lib is also inclueded and needs to be installed.
the result is cluster.txt